changed io method to specific header
