+++ include/StreamIO.h
@@ -1,0 +1,9 @@
+#ifndef WIND_STREAMIO_H
+#define WIND_STREAMIO_H
+
+#include "WindStream.h"
+
+// Prints the entire stream to stdout
+void WindStream_out(WindStream* wstream);
+
+#endif // WIND_STREAMIO_H
