+++ src/stream/WindStream.c
@@ -20,25 +20,3 @@ void WindStream_del(WindStream* wstream)
         }
         free(wstream);
 }
-
-
-
-
-
-
-void WindStream_out(WindStream* wstream)
-{
-        printf("[ ");
-        WindObject* objPtr = wstream->front;
-        while(objPtr != NULL)
-        {
-                switch(objPtr->type)
-                {
-                case WindType_None:
-                        printf("%s ", "None");
-                        break;
-                }
-                objPtr = objPtr->next;
-        }
-        puts("]");
-}
